% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART_Activate_Categories.r
\name{ART_Activate_Categories}
\alias{ART_Activate_Categories}
\title{ART_Activate_Categories}
\usage{
ART_Activate_Categories(input, weight, bias)
}
\arguments{
\item{input}{A vector of size NumFeatures that contains the input signal into the network}

\item{weight}{Matrix of size NumFeatures-by-NumCategories which holds the weights of the network}

\item{bias}{Constant that is used to differentiate between very similar category activation values}
}
\value{
Vector of size NumCategories that holds the activation value for each category
}
\description{
Activates the categories in an ART network
}
\details{
This function returns a vector of category activations for the given 
input vector, weight matrix, and bias value

The length of the INPUT vector must equal the number of rows in the WEIGHT matrix, 
and the BIAS value must be within the range [0, 1] (although values very near 0 are best).
}

