% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART_Calculate_Match.r
\name{ART_Calculate_Match}
\alias{ART_Calculate_Match}
\title{ART_Calculate_Match}
\usage{
ART_Calculate_Match(input, weightVector)
}
\arguments{
\item{input}{Vector of size NumFeatures that contains the input signal into the network}

\item{weightVector}{Matrix of size NumFeatures which holds the weights of the network for a given category}
}
\value{
Measure of the degree of match between the input and the current category
}
\description{
Calculates the match value of an input to a category.
}
\details{
This function returns a value which represents the amount of match
between the given input and the given category.

The length of the INPUT vector must equal the length of the WEIGHTVECTOR.
}

