% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART_Complement_Code.r
\name{ART_Complement_Code}
\alias{ART_Complement_Code}
\title{ART_Complement_Code}
\usage{
ART_Complement_Code(data)
}
\arguments{
\item{data}{Matrix of size NumFeatures-by-NumSamples that holds the data to be complement coded.}
}
\value{
Data that has been complement coded. It is a matrix of size 2*NumFeatures-by-NumSamples.
}
\description{
Complement codes the data for use with an ART network.
}
\details{
This function complement codes the given data where the complement of x is 1-x.
}

