% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART_Update_Weights.r
\name{ART_Update_Weights}
\alias{ART_Update_Weights}
\title{ART_Update_Weights}
\usage{
ART_Update_Weights(input, weight, categoryNumber, learningRate)
}
\arguments{
\item{input}{Vector of size NumFeatures that contains the input signal into the network.}

\item{weight}{Matrix of size NumFeatures-by-NumCategories which holds the weights of the network.}

\item{learningRate}{Rate at which the network should learn new inputs.}

\item{categoryNumer}{Number of the category that codes the current input.}
}
\value{
List containing the UPDATEDWEIGHT and WEIGHTCHANGE. The UPDATEDWEIGHT is a matrix of size NumFeatures-by-NumCategories
that holds the new weights of the network after the input has been successfully learned.
The WEIGHTCHANGE is a value (0 or 1) which relays whether or not
the weight matrix was changed during this updating. Here, 0 represents
no change and 1 represents a change.
}
\description{
Updates the weight matrix of an ART network.
}
\details{
This function returns a new weight matrix that has "learned" the input
in the given category, as well as a value correspoding to whether or 
not the weight matrix was changed (0 = no change; 1 = change).

The length of the INPUT vector must equal the number of rows in the WEIGHT matrix, the CATEGORYNUMBER must
be in the range [1, NumCategories], and the LEARNINGRATE must be in the range [0, 1].
}

