% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTMAP_Add_New_Category.r
\name{ARTMAP_Add_New_Category}
\alias{ARTMAP_Add_New_Category}
\title{ARTMAP_Add_New_Category}
\usage{
ARTMAP_Add_New_Category(weight, mapField)
}
\arguments{
\item{weight}{Matrix of size NumFeatures-by-NumCategories which holds the weights of the network.}

\item{mapField}{Matrix of size 1-by-NumCategories which holds the associations between categories and classes.}
}
\value{
Pair of values [RESIZEDWEIGHT, RESIZEDMAPFIELD]. The RESIZEDWEIGHT is a matrix of size NumFeatures-by-NumCategories+1
which holds the weights of the old matrix plus a new category of all
values of 1. The RESIZEDMAPFIELD is a matrix of size 1-by-NumCategories+1
which holds the given map field matrix except that it contains
one more element initialized to 0.
}
\description{
Adds a new category to the given weight matrix and map field.
}
\details{
This function returns a new weight matrix which is identical to the
given weight matrix except that it contains one more category which
is initialized to all 1's. It also returns a new map field matrix
which is identical to the given map field matrix except that it contains
one more element initialized to 0.
}

