% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTMAP_Classify.r
\name{ARTMAP_Classify}
\alias{ARTMAP_Classify}
\title{ARTMAP_Classify}
\usage{
ARTMAP_Classify(artmap_network, data)
}
\arguments{
\item{artmap_network}{The trained ARTMAP network. It should be created with ARTMAP_Create_Network() and trained with ARTMAP_Learn()}

\item{data}{Classification data to be presented to the network. It is a matrix of size NumFeatures-by-NumSamples.}
}
\value{
Vector of size NumSamples that holds the class in which the ARTMAP network placed each sample.
}
\description{
Uses an ARTMAP network to classify the given input data.
}
\details{
This function uses an ARTMAP network to classify the given input data with 
the specified vigilance parameter. Each sample of the data is presented to
the network, which classifies each sample. The function returns the 
classification of each sample. If the classification of the sample requires
that a new category be created, the class for that sample is set to -1.
}

