% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTMAP_Learn.r
\name{ARTMAP_Learn}
\alias{ARTMAP_Learn}
\title{ARTMAP_Learn}
\usage{
ARTMAP_Learn(artmap_network, data, supervisor)
}
\arguments{
\item{artmap_network}{ARTMAP network to be trained It should be created with ARTMAP_Create_Network().}

\item{data}{Training data to be presented to the network. It is a matrix of size NumFeatures-by-NumSamples.}

\item{supervisor}{the correct classification for each input vector. It is a matrix of size 1-by-NumSamples.}
}
\value{
New ARTMAP network which has learned the input data.
}
\description{
Trains an ARTMAP network on the given input data.
}
\details{
This function trains an ARTMAP network on the given input data. Each sample
of the data is presented to the network, which categorizes each sample
and compares that category's entry in the map field to the supervisor signal.
If the map field value and the supervisor signal do not match, match-tracking
is induced until a category is found to code the input correctly. This
category then learns the input vector.

The function returns a new ARTMAP network which has learned the input data
according to the supervisor signal. If the maximum number of categories 
is reached and an appropriate categorization of the input cannot be made,
no learning occurs. The program prints out a warning message that the 
maximum category limit has been reached and begins to process the next
input vector.
}

